+++ source/scripts/migrations/v2.1.0/migration.js
@@ -1,0 +1,8 @@
+const { DataTypes } = require("sequelize");
+const { connectToDatabase } = require("../../../../database.js");
+
+connectToDatabase("migration").then(async database => {
+	const queryInterface = database.getQueryInterface();
+	queryInterface.addColumn("Company", "bountyBoardOpenTagId", { type: DataTypes.STRING });
+	queryInterface.addColumn("Company", "bountyBoardCompletedTagId", { type: DataTypes.STRING });
+});